CRUD de productos casi listo

+++ src/Core/Domain/Entities/Categoria.cs
@@ -2,7 +2,7 @@ namespace src.Models.Domain;
 
 public class Categoria
 {
-    public int IdCategoria { get; init; }
+    public short IdCategoria { get; init; }
     public string Nombre { get; set; }
     public string Descripcion { get; set; }
     public Familia Familia { get; set; }
